requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
python-dotenv==1.0.0
cryptography==41.0.7
schedule==1.2.1
webdriver-manager==4.0.1
